containers/orchestration/requirements.txt
anyio>=4.4.0
typing_extensions>=4.8.0
pydantic==1.10.13
fastapi==0.112.2
uvicorn[standard]
httpx
pytest
lxml
frozendict
requests
python-dotenv
python-multipart
pytest-env
pytest-asyncio
testcontainers[compose]==4.1.1
websockets
asyncio
sqlalchemy
requests-mock
psycopg2-binary==2.9.9
pytest-mock
opentelemetry-distro==0.48b0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-prometheus==0.48b0
